Search api interaction (#12)
A Whole tonne of stuff went down. The EDI was made, then remade and then started again!

This branch really just kicks things off. there was an original EDI in the `edi` folder. this is now replaced by `the-edi`. 

There is services and components that mainly deal with communicating with the api server and then displaying the results on the screen.